Add dev option to install
This symlinks the source code to the destination
Just run: `./install --dev`.

Also, the uninstall script checks if the installation is a symlink and
removes accordingly

+++ uninstall.sh
@@ -3,18 +3,40 @@
 set -eE
 trap "_log --error 'Failed to uninstall monkeypatch'" ERR
 
-SOURCE_DIR="$(dirname ${BASH_SOURCE[0]})"
+SOURCE_DIR="$(realpath $(dirname ${BASH_SOURCE[0]}))"
 source $SOURCE_DIR/common.sh
 
 function rm_mondir() {
-    # Remove monkeypatsh executable and wrappers
-    rm -r $MON_DIR
-    _log "Removed monkeypatsh binary and wrappers from $MON_DIR and this directory itself as well as scripts"
+    # Remove monkeypatsh executable, wrappers and completions
+
+    local msg
+
+    # If we have symlinked (probably a dev install)
+    # we remove by parts
+    if [ -L "$(echo $MON_DIR | sed 's:/*$::')" ]; then
+        # Remove registered folder (not present in source)
+        rm -r "$MON_REGISTERED"
+        # Remove all completions besides the one for `mon` (`completions/mon` is present in source)
+        find "$MON_COMPLETIONS" -type f | grep -v "$MON_COMPLETIONS/mon"  | xargs -I {} rm {}
+        # Remove symlink to source
+        rm "$MON_DIR"
+
+        msg="Removed symlink to monkeypatsh executable, wrappers and completions."
+
+    # If not, just remove the copy
+    else
+        rm -r "$MON_DIR"
+        msg="Removed monkeypatsh executable, wrappers and completions."
+    fi
+
+    _log "$msg"
 }
 
 function rm_monrc_file() {
-    rm $MONRC_FILE
-    _log "Removed $MONRC_FILE"
+    if [ -f "$MONRC_FILE" ]; then
+        rm $MONRC_FILE
+        _log "Removed $MONRC_FILE"
+    fi
 }
 
 function rm_monconfig_file() {
